[IMP] Use of subdirectories in src/pages
