+++ source/app/blueprints/pages/case/templates/case.html
@@ -328,7 +328,7 @@
 
 <script src="/static/assets/js/plugin/showdown/showdown.min.js"></script>
 <script src="/static/assets/js/iris/datatablesUtils.js"></script>
-<script type="module" src="/static/assets/js/iris/case.js"></script>
+<script src="/static/assets/js/iris/case.js"></script>
 <script src="/static/assets/js/iris/manage.cases.common.js"></script>
 <script src="/static/assets/js/iris/case.common.js"></script>
 <script src="/static/assets/js/iris/case.summary.js"></script>
