+++ source/app/blueprints/pages/dashboard/templates/index.html
@@ -284,7 +284,7 @@
     <script src="/static/assets/js/plugin/datatables/dataTables.select.min.js"></script>
     <script src="/static/assets/js/plugin/datatables/buttons.html5.min.js"></script>
     <script src="/static/assets/js/plugin/datatables/buttons.print.min.js"></script>
-    <script type="module" src="/static/assets/js/iris/dashboard.js"></script>
+    <script src="/static/assets/js/iris/dashboard.js"></script>
     <script src="/static/assets/js/core/charts.js"></script>
     <script src="/static/assets/js/core/Chart.PieceLabel.min.js"></script>
