--- ui/src/pages/demo.svelte
@@ -1,3 +1,0 @@
-<script>
-    import '../css/demo.css';
-</script>
