+++ ui/src/pages/demo/demo.svelte
@@ -1,0 +1,3 @@
+<script>
+    import '../../css/demo.css';
+</script>
