+++ ui/src/pages/demo/index.js
@@ -1,0 +1,7 @@
+import Demo from "./demo.svelte";
+
+const app = new Demo({
+    target: document.body,
+});
+
+export default app;
