+++ ui/src/pages/login/index.js
@@ -1,0 +1,8 @@
+import Login from "./login.svelte"
+
+
+const app = new Login({
+    target: document.body,
+});
+
+export default app;
